fix: improve page context serialization in state (#6389)
closes #6388
