+++ projects/core/src/cms/store/selectors/component.selectors.spec.ts
@@ -4,7 +4,7 @@ import { select, Store, StoreModule } from '@ngrx/store';
 import { CmsComponent, PageType } from '../../../model/cms.model';
 import { PageContext } from '../../../routing/index';
 import { initialLoaderState, LoaderState } from '../../../state/index';
-import { serializePageContext } from '../../../util/serialization-utils';
+import { serializePageContext } from '../../utils/cms-utils';
 import { CmsActions } from '../actions/index';
 import { ComponentsContext, StateWithCms } from '../cms-state';
 import * as fromReducers from '../reducers/index';
@@ -102,7 +102,7 @@ describe('Cms Component Selectors', () => {
           )
         );
 
-        const serializedPageContext = serializePageContext(pageContext);
+        const serializedPageContext = serializePageContext(pageContext, true);
         expect(result.component).toEqual({
           uid: componentUid,
           typeCode: component.typeCode,
@@ -181,7 +181,7 @@ describe('Cms Component Selectors', () => {
             id: 'xxx',
             type: PageType.CONTENT_PAGE,
           };
-          const serializedPageContext = serializePageContext(pageContext);
+          const serializedPageContext = serializePageContext(pageContext, true);
           store.dispatch(
             new CmsActions.LoadCmsComponentSuccess(
               component,
@@ -255,7 +255,7 @@ describe('Cms Component Selectors', () => {
           id: 'xxx',
           type: PageType.CONTENT_PAGE,
         };
-        const serializedPageContext = serializePageContext(pageContext);
+        const serializedPageContext = serializePageContext(pageContext, true);
 
         store.dispatch(
           new CmsActions.LoadCmsComponentSuccess(
@@ -356,7 +356,7 @@ describe('Cms Component Selectors', () => {
           id: 'xxx',
           type: PageType.CONTENT_PAGE,
         };
-        const serializedPageContext = serializePageContext(pageContext);
+        const serializedPageContext = serializePageContext(pageContext, true);
 
         store.dispatch(
           new CmsActions.LoadCmsComponentSuccess(
